# sirius_rest_service/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SIRIUS Nightsky API",
    "description": "REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6.0.0-SNAPSHOT",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/{projectId}/jobs/run-command": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Start computation for given command and input.",
        "description": "Start computation for given command and input.",
        "operationId": "startCommand",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to perform the command for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/JobOptField"
              },
              "default": [
                "progress"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "the command and the input to be executed",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Job of the command to be executed.\n\n DEPRECATED: this endpoint is based on local file paths and will likely be removed in future versions of this API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/api/projects/{projectId}/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get List of all available jobs with information such as current state and progress (if available).",
        "description": "Get List of all available jobs with information such as current state and progress (if available).",
        "operationId": "getJobs",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to run jobs on",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/JobOptField"
              },
              "default": []
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Start computation for given compounds and with given parameters.",
        "description": "Start computation for given compounds and with given parameters.",
        "operationId": "startJob",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to run jobs on",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/JobOptField"
              },
              "default": [
                "command",
                "progress"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "configuration of the job that will be submitted of the job to be returned",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "* Delete ALL jobs.",
        "description": "* Delete ALL jobs. Specify how to behave for running jobs.",
        "operationId": "deleteJobs",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to delete jobs from",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cancelIfRunning",
            "in": "query",
            "description": "If true job will be canceled if it is not finished. Otherwise,\n                        deletion will fail for running jobs or request will block until job has finished.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "awaitDeletion",
            "in": "query",
            "description": "If true request will block until deletion succeeded or failed.\n                        If the job is still running the request will wait until the job has finished.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/projects/{projectId}/jobs/from-config": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Start computation for given compounds and with parameters from a stored job-config.",
        "description": "Start computation for given compounds and with parameters from a stored job-config.",
        "operationId": "startJobFromConfig",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to run jobs on",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobConfigName",
            "in": "query",
            "description": "name if the config to be used",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recompute",
            "in": "query",
            "description": "enable or disable recompute. If null the stored value will be used.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/JobOptField"
              },
              "default": [
                "command",
                "progress"
              ]
            }
          }
        ],
        "requestBody": {
          "description": "compound ids to be computed",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/job-configs/{name}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Request job configuration with given name.",
        "description": "Request job configuration with given name.",
        "operationId": "getJobConfig",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the job-config to return",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeConfigMap",
            "in": "query",
            "description": "if true the generic configmap will be part of the output",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{@link JobSubmission JobSubmission} for given name.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmission"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Add new job configuration with given name.",
        "description": "Add new job configuration with given name.",
        "operationId": "postJobConfig",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the job-config to add",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "overrideExisting",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "description": "to add",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Probably modified name of the config (to ensure filesystem path compatibility).",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete job configuration with given name.",
        "description": "Delete job configuration with given name.",
        "operationId": "deleteJobConfig",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "name of the job-config to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/projects/{projectId}/jobs/{jobId}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job information and its current state and progress (if available).",
        "description": "Get job information and its current state and progress (if available).",
        "operationId": "getJob",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to run jobs on",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "in": "path",
            "description": "of the job to be returned",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/JobOptField"
              },
              "default": [
                "progress"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete job.",
        "description": "Delete job. Specify how to behave for running jobs.",
        "operationId": "deleteJob",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to delete job from",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "in": "path",
            "description": "of the job to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cancelIfRunning",
            "in": "query",
            "description": "If true job will be canceled if it is not finished. Otherwise,\n                        deletion will fail for running jobs or request will block until job has finished.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "awaitDeletion",
            "in": "query",
            "description": "If true request will block until deletion succeeded or failed.\n                        If the job is still running the request will wait until the job has finished.",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
...
    "/api/projects/{projectId}/has-jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "operationId": "hasJobs",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeFinished",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandSubmission": {
        "required": [
          "command"
        ],
        "type": "object",
        "properties": {
          "compoundIds": {
            "type": "array",
            "description": "Compounds that should be the input for this Job\n Will be converted to the respective alignedFeatureIds for computation.\n\n At least one compoundId or alignedFeatureId needs to be specified.",
            "nullable": true,
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "alignedFeatureIds": {
            "type": "array",
            "description": "Features (aligned over runs) that should be the input for this Job\n\n At least one compoundId or alignedFeatureId needs to be specified.",
            "nullable": true,
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inputPaths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "JobOptField": {
        "type": "string",
        "nullable": true,
        "enum": [
          "none",
          "command",
          "progress",
          "affectedIds"
        ]
      },
...
    }
  }
}